[FIX] 매물 리스트/상세내역 조회 로직 수정, 청약 신청 로직 수정

+++ woorepie/src/main/java/com/piehouse/woorepie/estate/dto/RedisEstatePrice.java
@@ -3,6 +3,8 @@ package com.piehouse.woorepie.estate.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class RedisEstatePrice {
@@ -13,6 +15,6 @@ public class RedisEstatePrice {
 
     private Integer tokenAmount;
 
-    private Integer dividendYield;
+    private BigDecimal dividendYield;
 
 }
